requirements.txt
pip==23.2
chainlit==0.5.2
langchain==0.0.233
numpy==1.25.0
openai==0.27.8
pandas==1.5.3
Pillow==9.5.0
Requests==2.31.0
scikit_learn==1.3.0
scipy==1.11.1
streamlit==1.24.0
numpy-stl==3.0.1
pygltflib==1.15.6
scikit-image==0.21.0
pytesseract==0.21.0